+++ module/actor/actor-covenant-sheet.js
@@ -37,7 +37,7 @@ export class ArM5eCovenantActorSheet extends ArM5eActorSheet {
     // editable, the items array, and the effects array.
     const context = super.getData();
 
-    context.metadata = CONFIG.ARM5E;
+    context.config = CONFIG.ARM5E;
     log(false, "Covenant-sheet getData");
     log(false, context);
     context.data.world = {};