+++ module/actor/actor-laboratory-sheet.js
@@ -28,7 +28,7 @@ export class ArM5eLaboratoryActorSheet extends ArM5eActorSheet {
   getData() {
     const context = super.getData();
 
-    context.metadata = CONFIG.ARM5E;
+    context.config = CONFIG.ARM5E;
 
     context.data.world = {};
     context.data.world.covenants = game.actors