+++ module/actor/actor-sheet.js
@@ -5,7 +5,7 @@
 
 import { simpleDie, stressDie } from "../dice.js";
 import { resetOwnerFields } from "../item/item-converter.js";
-import { ARM5E } from "../metadata.js";
+import { ARM5E } from "../config.js";
 import { log, getLastMessageByHeader, calculateWound, getDataset } from "../tools.js";
 import ArM5eActiveEffect from "../helpers/active-effects.js";
 import { VOICE_AND_GESTURES_VALUES } from "../constants/voiceAndGestures.js";
@@ -121,8 +121,8 @@ export class ArM5eActorSheet extends ActorSheet {
     context.data = actorData.data;
     context.flags = actorData.flags;
 
-    context.metadata = CONFIG.ARM5E;
-    context.metadata.constants = { VOICE_AND_GESTURES_VALUES: VOICE_AND_GESTURES_VALUES };
+    context.config = CONFIG.ARM5E;
+    context.config.constants = { VOICE_AND_GESTURES_VALUES: VOICE_AND_GESTURES_VALUES };
 
     context.data.dtypes = ["String", "Number", "Boolean"];