+++ module/helpers/voiceAndGestures.js
@@ -64,9 +64,11 @@ function findVoiceAndGesturesActiveEffects(effects) {
   const actualVoiceEffect = ArM5eActiveEffect.findFirstActiveEffectBySubtype(effects, VOICE);
   const actualGesturesEffect = ArM5eActiveEffect.findFirstActiveEffectBySubtype(effects, GESTURES);
   return {
-    // only one change for voice and gestures => index 0 hardcoded
+    // // only one change for voice and gestures => index 0 hardcoded
     voice: actualVoiceEffect?.getFlag("arm5e", "value")[0] || DEFAULT_VOICE,
     gestures: actualGesturesEffect?.getFlag("arm5e", "value")[0] || DEFAULT_GESTURES
+    // voice: DEFAULT_VOICE,
+    // gestures: DEFAULT_GESTURES
   };
 }